Configure the CD for Maven and NPM (#25)
The CI runs on a build matrix going through each individual vocabulary bundle.

Maven artifacts are published on push event, which makes sure that the -SNAPSHOT artifact is always up-to-date with `master`. NPM modules are only published on release, because due to the current setup, deploying review apps would be complex/clunky.